+++ _networkMocks/report/mockVariants/feature-testing/missing1.ts
@@ -43,6 +43,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           type: 'Select',
           staticOptions: [
@@ -63,6 +64,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           type: 'daterange',
           defaultValue: '2003-02-01 - 2006-05-04',
@@ -80,6 +82,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         mandatory: false,
         filter: {
           type: 'autocomplete',
@@ -106,6 +109,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           type: 'autocomplete',
           dynamicOptions: {
@@ -124,6 +128,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           mandatory: false,
           type: 'text',
@@ -139,6 +144,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           mandatory: false,
           type: 'date',
@@ -157,6 +163,7 @@ export const featureTestingMissing1: components['schemas']['VariantDefinition'] 
         calculated: false,
         header: false,
         defaultsort: false,
+        fieldSource: 'specfield',
         filter: {
           mandatory: false,
           type: 'multiselect',