+++ src/dpr/utils/Personalisation/personalisationUtils.ts
@@ -10,6 +10,7 @@ import { FilterValue, FilterValueWithOptions } from '../../components/_filters/t
 import { defaultFilterValue } from './types'
 import { FiltersType } from '../../components/_filters/filtersTypeEnum'
 import { uiDateToApi } from '../dateHelper'
+import { logger } from '../logger'
 
 export const saveDefaults = async (type: FiltersType, res: Response, req: Request, services: Services) => {
   const { dprUser } = localsHelper.getValues(res)
@@ -106,18 +107,44 @@ export const setUserContextDefaults = (res: Response, filters: FilterValue[]) =>
   const { dprUser } = localsHelper.getValues(res)
   const { activeCaseLoadId } = dprUser
 
-  filters.forEach((filter) => {
+  logger.info('PERSONALISATION DEBUG: SETTING USER CONTEXT DEFAULTS')
+  logger.info(`PERSONALISATION DEBUG: dprUser: ${JSON.stringify(dprUser)}`)
+  logger.info(`PERSONALISATION DEBUG: activeCaseLoadId: ${JSON.stringify(activeCaseLoadId)}`)
+
+  filters.forEach((filter, index) => {
+    logger.info(
+      `PERSONALISATION DEBUG: (${index}) : Is autocomplete: `,
+      filter.type.toLocaleLowerCase() === FilterType.autocomplete.toLocaleLowerCase(),
+    )
+    logger.info(
+      `PERSONALISATION DEBUG: (${index}) : Is establishment: `,
+      filter.text.toLocaleLowerCase().includes('establishment'),
+    )
+    logger.info(
+      `PERSONALISATION DEBUG: (${index}) : Has activeCaseLoadId: `,
+      activeCaseLoadId && activeCaseLoadId.length,
+    )
+
     if (
       filter.type.toLocaleLowerCase() === FilterType.autocomplete.toLocaleLowerCase() &&
       filter.text.toLocaleLowerCase().includes('establishment') &&
-      activeCaseLoadId?.length
+      activeCaseLoadId &&
+      activeCaseLoadId.length
     ) {
+      logger.info(`PERSONALISATION DEBUG: (${index}): Autocomplete filter found:`)
+      logger.info(`PERSONALISATION DEBUG: (${index}): updated filter: ${JSON.stringify(filter)}`)
+
       const f = <FilterValueWithOptions>filter
       const option = f.options.find((opt) => opt.value === activeCaseLoadId)
 
       if (option) {
+        logger.info(`PERSONALISATION DEBUG: (${index}): Autocomplete option found:`)
+        logger.info(`PERSONALISATION DEBUG: (${index}): option: ${JSON.stringify(option)}`)
         f.value = option.text
-        f.staticOptionNameValue = activeCaseLoadId
+        f.staticOptionNameValue = option.value
+
+        logger.info(`PERSONALISATION DEBUG: (${index}): Updated filter:`)
+        logger.info(`PERSONALISATION DEBUG: (${index}): updated filter: ${JSON.stringify(f)}`)
       }
     }
   })